move profiles schemas to core for consistency
